Add transparent option, to only show real compiler messages
By default, distcc will show verbose messages when failing to distribute.
It will do this by itself, and even if the compilation also fails locally.

Failing to distribute is not really an error (in distcc), it might just be
that there are no slots available... Or perhaps the code does not compile.